Move calc_hydraulic_properties and enable extra_layer by default
